requirements.txt
# Core ML and Scientific Computing
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0
accelerate>=0.20.0
datasets>=2.12.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Configuration and Logging
pyyaml>=6.0
hydra-core>=1.3.0
wandb>=0.15.0
tensorboard>=2.13.0

# Data Storage and Processing
h5py>=3.8.0
zarr>=2.14.0
tqdm>=4.65.0

# NLP and Concept Processing
nltk>=3.8.0
wordnet>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Statistical Analysis
statsmodels>=0.14.0
pingouin>=0.5.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# GPU Monitoring (for RunPod)
gpustat>=1.1.0
nvidia-ml-py>=12.535.0

# Optional: Model optimization
bitsandbytes>=0.41.0
flash-attn>=2.0.0

# Optional: Advanced visualization
umap-learn>=0.5.3
bokeh>=3.2.0